+++ chen_cppcguilib/utils/string.cpp
@@ -8,7 +8,7 @@ static const std::string defaultColor = "\033[0m";
 string::string() {}
 string::string(const char* in) : str(in) { calculateVisibleLength(); }
 string::string(std::string_view in) : str(in) { calculateVisibleLength(); }
-string::string(int count, char c) : str(count, c) { calculateVisibleLength(); }
+string::string(size_t count, char c) : str(count, c) { calculateVisibleLength(); }
 
 size_t string::size() const { return str.size(); }
 size_t string::length() const { return visibleLength; }
@@ -89,7 +89,7 @@ const char* cgui::string::data() const
     return str.data();
 }
 
-int cgui::string::pushBackPos() const
+size_t cgui::string::pushBackPos() const
 {
     if (colorCount == 0) {
         return str.end() - str.begin();
